Moving CollaboratorsFragment to its package (repodetails/contributorslist)
