+++ app/src/main/java/br/com/luisfernandez/github/client/repodetails/contributorslist/CollaboratorsFragment.kt
@@ -1,4 +1,4 @@
-package br.com.luisfernandez.github.client.repodetails
+package br.com.luisfernandez.github.client.repodetails.contributorslist
 
 
 import android.os.Bundle
